requirements.txt
PySide6>=6.5
websocket-client>=1.5
